+++ tests/acceptance/test_pebbledoc.py
@@ -369,6 +369,27 @@ def test_pebbledoc_no_full_toc_name_no_main_module_header(
     assert exit_code == 0
 
 
+def test_pebbledoc_no_preserve_linewraps(
+    patch_open: Mock, patch_config_discovery: None
+) -> None:
+    """Test pebbledoc with option to remove linewraps."""
+    input_file = (
+        Path(__file__).parent / "expected" / "no_preserve_linewraps.md"
+    )
+    with open(input_file, "r") as f:
+        expected = f.read()
+
+    # create a run config and execute the code
+    namespace = utils.prepare_namespace(
+        source_directory=str(Path(__file__).parent / "resources"),
+        no_preserve_linewraps=True,
+    )
+    exit_code = cli._handle_args(namespace)
+
+    assert_write_call(patch_open, namespace.output, expected + "\n")
+    assert exit_code == 0
+
+
 # == SPECIAL TESTS =====================================================