+++ tests/utils.py
@@ -20,6 +20,7 @@ def prepare_namespace(
     no_collapsible_params: bool = False,
     no_references: bool = False,
     no_full_toc_name: bool = False,
+    no_preserve_linewraps: bool = False,
 ) -> argparse.Namespace:
     """
     Create a namespace with the attributes that pebbledoc expects.